fix: Preserve dtype in Constant mean function (#523)

+++ gpjax/mean_functions.py
@@ -147,9 +147,9 @@ class Constant(AbstractMeanFunction):
             Float[Array, "1"]: The evaluated mean function.
         """
         if isinstance(self.constant, Parameter):
-            return jnp.ones((x.shape[0], 1)) * self.constant.value
+            return jnp.ones((x.shape[0], 1), dtype=x.dtype) * self.constant.value
         else:
-            return jnp.ones((x.shape[0], 1)) * self.constant
+            return jnp.ones((x.shape[0], 1), dtype=x.dtype) * self.constant
 
 
 class Zero(Constant):
